test: add email code send test and fix mypy
