+++ apps/users/serializers/auth.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
 
 
-class EmailCodeSendRequestSerializer(serializers.Serializer):
+class EmailCodeSendRequestSerializer(serializers.Serializer): # type: ignore[type-arg]
     email = serializers.EmailField(required=True)
